# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Meridian Project Management API
  description: |
    Comprehensive REST API for the Meridian project management and team collaboration platform.
    
    ## Features
    - **User Management**: Authentication, user profiles, and session management
    - **Project Management**: Projects, tasks, milestones, and workflows
    - **Real-time Communication**: Messages, channels, and notifications
    - **Analytics**: Performance metrics, reporting, and insights
    - **Database Optimization**: Query analysis, caching, and performance monitoring
    - **Logging**: Structured logging, monitoring, and log management
    
    ## Authentication
    This API uses session-based authentication with secure tokens.
    Include the session token in the `Authorization` header as a Bearer token.
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse. Check response headers for current limits.
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with structured error responses.
  version: 2.0.0
  contact:
    name: Meridian Development Team
    url: https://github.com/meridian-project
    email: support@meridian.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3008
    description: Development server
  - url: https://api.meridian.com
    description: Production server
security:
  - BearerAuth: []
  - SessionAuth: []
paths:
...
  /api/db/cache:
    get:
      tags: [Database]
      summary: Cache performance and statistics
      description: Get cache performance metrics and health information
      operationId: getCacheStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cache performance statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Database]
      summary: Clear cache entries
      description: Clear cache entries with optional pattern and tag filtering
      operationId: clearCache
      security:
        - BearerAuth: []
      parameters:
        - name: pattern
          in: query
          description: Regex pattern to match cache keys
          schema:
            type: string
            example: "user:*"
        - name: tags
          in: query
          description: Comma-separated list of tags to filter
          schema:
            type: string
            example: "user,session"
      responses:
        '200':
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheClearResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/logs/search:
    get:
      tags: [Logging]
      summary: Search and filter logs
      description: Search logs with advanced filtering and pagination
      operationId: searchLogs
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query string
          schema:
            type: string
            example: "failed authentication"
        - name: level
          in: query
          description: Comma-separated log levels
          schema:
            type: string
            example: "error,warn"
        - name: category
          in: query
          description: Comma-separated categories
          schema:
            type: string
            example: "AUTH,API"
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: timeWindow
          in: query
          description: Time window in milliseconds
          schema:
            type: integer
            default: 3600000
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogSearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication
    SessionAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session-based authentication cookie
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request parameters"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          additionalProperties: true
      required: [error]

    # Health and System
...
    CacheStatsResponse:
      type: object
      properties:
        statistics:
          type: object
          properties:
            hits:
              type: integer
            misses:
              type: integer
            size:
              type: integer
            hitRatio:
              type: number
        health:
          type: object
          properties:
            status:
              type: string
              enum: [healthy, degraded, unhealthy]
            performance:
              type: object
              properties:
                hitRatio:
                  type: number
                avgResponseTime:
                  type: number
                memoryUsage:
                  type: number
        recommendations:
          type: array
          items:
            type: string
      required: [statistics, health]
    CacheClearResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        clearedCount:
          type: integer
          example: 42
        message:
          type: string
          example: "Cleared 42 cache entries"
      required: [success, clearedCount, message]
...
    LogSearchResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: integer
                format: int64
              level:
                type: string
              category:
                type: string
              message:
                type: string
              metadata:
                type: object
                additionalProperties: true
        total:
          type: integer
        query:
          type: string
        filters:
          type: object
          properties:
            level:
              type: array
              items:
                type: string
            category:
              type: array
              items:
                type: string
            userId:
              type: string
        timeWindow:
          type: integer
        limit:
          type: integer
      required: [logs, total]
...
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Insufficient permissions"
            code: "FORBIDDEN"
...